sequential: use os.OpenFile with O_FILE_FLAG_SEQUENTIAL_SCAN on Go 1.26+
Go 1.26 adds support for passing Windows file flags via os.OpenFile,
eliminating the need to call windows.CreateFile directly. Split the
implementation into build-tagged files: go1.26+ uses os.OpenFile
with the new O_FILE_FLAG_SEQUENTIAL_SCAN open flag, while older
versions retain the manual CreateFile approach.

Also bump golang.org/x/sys to v0.37.0 (provides the new constant)
and update the minimum Go version to 1.24.
