+++ sequential/go.mod
@@ -1,5 +1,5 @@
 module github.com/moby/sys/sequential
 
-go 1.18
+go 1.24.0
 
-require golang.org/x/sys v0.1.0
+require golang.org/x/sys v0.37.0
